requirements.txt
fake-bpy-module-latest
supabase